Move TernarySearchTree to own file
This class is taking up most of `map.ts`. I think it's time to split it out into its own file
