+++ src/vs/platform/workspace/common/workspace.ts
@@ -6,7 +6,7 @@
 import { localize } from 'vs/nls';
 import { Event } from 'vs/base/common/event';
 import { basename, extname } from 'vs/base/common/path';
-import { TernarySearchTree } from 'vs/base/common/map';
+import { TernarySearchTree } from 'vs/base/common/ternarySearchTree';
 import { extname as resourceExtname, basenameOrAuthority, joinPath, extUriBiasedIgnorePathCase } from 'vs/base/common/resources';
 import { URI, UriComponents } from 'vs/base/common/uri';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
